+++ ReMarket-02-08/api/src/main/java/ru/rrenat358/api/core/ProductDto.java
@@ -4,13 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 //@Data
 //@NoArgsConstructor
 //@AllArgsConstructor
 public class ProductDto {
     private Long id;
     private String title;
-    private Integer price;
+    private BigDecimal price;
 
 
     public Long getId() {
@@ -29,18 +31,18 @@ public class ProductDto {
         this.title = title;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
     public ProductDto() {
     }
 
-    public ProductDto(Long id, String title, Integer price) {
+    public ProductDto(Long id, String title, BigDecimal price) {
         this.id = id;
         this.title = title;
         this.price = price;
